+++ client/src/components/commentsection/CommentSection.js
@@ -1,10 +1,9 @@
 import React, { useState } from "react";
 import Rating from "react-rating-stars-component";
 import axios from "axios";
-// import { Clouds } from "../twinklingbackground/TwinklingBackground";
-import commentsDesktopImage from '../../images/commentsDesktop5.png';
+import commentsDesktopImage from '../../images/commentsDesktop.png';
 import commentsTabletImage from '../../images/commentsTablet.png';
-import commentsMobileImage from '../../images/commentsMobile1.png'; 
+import commentsMobileImage from '../../images/commentsMobile.png'; 
 import Comments from "../../images/placeholder3.png";
 import TwinklingBackground from '../twinklingbackground/TwinklingBackground'
 import { BsSendFill } from "react-icons/bs";
@@ -23,15 +22,7 @@ import {
   Subtitle,
   TextArea,
   CommentBtnWrapper,
-CommentCloud1,
-CommentCloud2,
-CommentCloud3,
-CommentCloud4,
-CommentCloud5,
-CommentCloud6,
-} from "./CommentSectionElements"; // You'll need to define these styled components
-// import StarBackground2 from '../starbackground2/StarBackground2';
-// import CommentsTwinkling from "./CommentsTwinkling"
+} from "./CommentSectionElements";
 
 const CommentSection = () => {
   const [rating, setRating] = useState(0);
@@ -63,24 +54,14 @@ const CommentSection = () => {
     }
   };
 
-  // const backgroundImage = commentsDesktop;
-
-
   return (
     <CommentSectionContainer id="comments">
       <TwinklingBackground />
-      {/* <Clouds style={{zIndex: '10'}} /> */}
       <CommentsBg
         backgroundImageDesktop={commentsDesktopImage}
         backgroundImageTablet={commentsTabletImage}
         backgroundImageMobile={commentsMobileImage}
       />
-        {/* <CommentCloud1 style={{ zIndex: '10' }} />
-        <CommentCloud2 style={{ zIndex: '10' }} /> */}
-        {/* <CommentCloud3 style={{ zIndex: '10' }} /> */}
-        {/* <CommentCloud4 style={{ zIndex: '10' }} /> */}
-        {/* <CommentCloud5 style={{ zIndex: '10' }} /> */}
-        {/* <CommentCloud6 style={{ zIndex: '10' }} /> */}
       <CommentWrapper style={{zIndex: '20'}} onSubmit={handleCommentSubmit}>
         <CommentRow>
         <ImgWrap>
@@ -124,7 +105,6 @@ const CommentSection = () => {
               </CommentBtnWrapper>
             </FormWrap>
           </TextWrapper>
-
         </CommentRow>
       </CommentWrapper>
     </CommentSectionContainer>
