+++ client/src/components/hero/Hero.js
@@ -12,13 +12,6 @@ import {
   HeroBtnWrapper,
   ArrowFilled,
   ArrowCircle,
-  HeroCloud1,
-  HeroCloud2,
-  HeroCloud3,
-  HeroCloud4,
-  HeroCloud5,
-  HeroCloud6,
-  // HeroCloud7,
 } from "./HeroElements";
 
 const HeroSection = () => {
@@ -38,13 +31,6 @@ const HeroSection = () => {
           cometImage="https://i.imgur.com/kW59Usj.png"
           earthImage="https://i.imgur.com/qTwsepd.png"
         />
-        <HeroCloud1 style={{ zIndex: '10' }} />
-        <HeroCloud2 style={{ zIndex: '10' }} />
-        <HeroCloud3 style={{ zIndex: '10' }} />
-        <HeroCloud4 style={{ zIndex: '10' }} />
-        <HeroCloud5 style={{ zIndex: '10' }} />
-        <HeroCloud6 style={{ zIndex: '10' }} />
-        {/* <HeroCloud7 style={{ zIndex: '10' }} /> */}
       </HeroBg>
       <HeroContent style={{zIndex: '20'}}>
         <StylishText />
