+++ client/src/components/hero/HeroElements.js
@@ -1,36 +1,8 @@
-import styled, { keyframes } from "styled-components";
+import styled from "styled-components";
 import { FaRocket, FaArrowAltCircleDown } from "react-icons/fa";
-import heroCloud1 from "../../images/heroCloud1.png";
-import heroCloud2 from "../../images/heroCloud2.png";
-import heroCloud3 from "../../images/heroCloud3.png";
-import Clouds from "../../images/clouds2.png";
-import Clouds2 from "../../images/clouds2reverse.png";
 
 
 const maxWidth = "700px"; // Common max-width for various components
-
-const moveCloud = keyframes`
-  /* 0%
-  {
-    opacity:0;
-    transform: translateY(0);
-  } 
-  10%, 90% 
-  {
-    opacity: 1;
-  }
-  {
-    opacity: 100%;
-    transform: translateY(-100px);
-  }  */
-  
-  from {
-    left: 100%; /* Start off-screen from the right */
-  }
-  to {
-    left: -100%; /* End off-screen on the left */
-  }
-`;
 
 // Hero Container
 export const HeroContainer = styled.div`
@@ -236,105 +208,3 @@ export const ArrowCircle = styled(FaArrowAltCircleDown)`
   margin-left: 8px;
   font-size: 20px;
 `;
-
-export const HeroCloud1 = styled.div`
-  background: url(${Clouds}) center center;
-  background-size: contain;
-  width: 80%;
-  height: 80%;
-  position: absolute;
-  top: 3%;
-  left: 100%;
-  animation: ${moveCloud} 60s linear infinite;
-  background-repeat: no-repeat;
-  /* filter: brightness(0.5) blur(3px); */
-`;
-
-export const HeroCloud2 = styled.div`
-  background: url(${Clouds2}) no-repeat;
-  position: absolute;
-  width: 700px;
-  height: 700px;
-  top: 20%;
-  left: -100%;
-  background-size: contain;
-  animation: ${moveCloud} 50s linear infinite;
-  /* animation-delay: 2s; */
-  /* filter: brightness(0.2) blur(3px); */
-`;
-
-export const HeroCloud3 = styled.div`
-  background: url(${Clouds}) no-repeat;
-  position: absolute;
-  top: 30%;
-  left: 0;
-  background-size: contain;
-  width: 80%;
-  height: 80%;
-  animation: ${moveCloud} 120s linear infinite;
-  /* animation-delay: 3s; */
-
-  @media screen and (max-width: 480px) {
-    display: none;
-  }
-`;
-
-export const HeroCloud4 = styled.div`
-  background: url(${Clouds}) no-repeat;
-  background-size: contain;
-  width: 50%;
-  height: 50%;
-  position: absolute;
-  top: 300px;
-  left: 0;
-  animation: ${moveCloud} 70s linear infinite;
-  background-repeat: no-repeat;
-  animation-delay: -20s;
-`;
-
-export const HeroCloud5 = styled.div`
-  background: url(${Clouds2}) no-repeat;
-  position: absolute;
-  top: 150px;
-  left: 0;
-  background-size: contain;
-  width: 60%;
-  height: 60%;
-  animation: ${moveCloud} 90s linear infinite;
-  animation-delay: -5s;
-`;
-
-export const HeroCloud6 = styled.div`
-  background: url(${Clouds}) no-repeat;
-  position: absolute;
-  /* top: 15%; */
-  top: 75px;
-  left: 0;
-  background-size: contain;
-  width: 900px;
-  height: 350px;
-  animation: ${moveCloud} 90s linear infinite;
-  animation-delay: -10s;
-
-  @media screen and (max-width: 480px) {
-    display: none;
-  }
-`;
-
-// export const HeroCloud7 = styled.div`
-//   background: url(${heroCloud7}) no-repeat;
-//   position: absolute;
-//   /* top: 15%; */
-//   top: 75px;
-//   left: 0;
-//   background-size: contain;
-//   width: 400px;
-//   height: 350px;
-//   animation: ${moveCloud} 90s linear infinite;
-//   animation-delay: -10s;
-//   filter: brightness(0.25) blur(3px);
-
-//   @media screen and (max-width: 480px) {
-//     display: none;
-//   }
-// `;
