--- client/src/components/stylishText/StylishText.css
@@ -1,49 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@900&display=swap');
-
-/* StylishText component styles */
-.stylishText {
-  position: relative;
-  font-family: 'Roboto', sans-serif; /* Changed to Roboto for a more similar look */
-  font-size: 150px; /* Adjustable to your preference */
-  font-weight: 900; /* Maximum weight for robustness */
-  color: #D6A2E8; /* Base color for the text */
-  background: linear-gradient(145deg, #E6E6E6, #FFD7FF); /* Gradient to mimic the light reflection */
-  background-clip: text;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  margin: 0 auto;
-  text-align: center;
-  user-select: none;
-
-  /* Adding multiple layers of text-shadow for the 3D effect */
-  text-shadow:
-    1px 1px 0 #EBB4FF,
-    2px 2px 0 #DAA2FF,
-    3px 3px 0 #C690FF,
-    4px 4px 0 #B27EFF,
-    5px 5px 0 #9E6CFF,
-    6px 6px 0 #8A5AFF,
-    7px 7px 0 #7648FF,
-    8px 8px 0 #6236FF,
-    9px 9px 0 #4E24FF,
-    10px 10px 10px rgba(0, 0, 0, 0.4); /* Outer dark shadow for depth */
-}
-
-/* Media queries for responsive font sizing */
-@media (max-width: 1024px) {
-  .stylishText {
-    font-size: 100px;
-  }
-}
-
-@media (max-width: 768px) {
-  .stylishText {
-    font-size: 72px;
-  }
-}
-
-@media (max-width: 480px) {
-  .stylishText {
-    font-size: 48px;
-  }
-}
