+++ server/.env.example
@@ -1,0 +1,9 @@
+PORT=
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+SESSION_SECRET=
+MONGO_URI=
+NODE_ENV=
+REDIRECT_URL=
+JWT_SECRET_KEY=
+FRONTEND_URL=
